python-data-dictionary-to-csv/requirements.txt
tableauserverclient==0.23
python-dotenv==0.21.0